Split helper file into multiple smaller ones
