+++ packages/dear-imgui/cpp/CMakeLists.txt
@@ -95,6 +95,10 @@ if (DEFINED EMSCRIPTEN)
 	add_executable(reactDearImgui
         src/main.cpp
         src/shared.cpp
+        src/color_helpers.cpp
+        src/imgui_helpers.cpp
+        src/webgpu_helpers.cpp
+        src/yoga_helpers.cpp
         src/element/layout_node.cpp
         src/element/element.cpp
         src/widget/widget.cpp
