+++ packages/dear-imgui/cpp/include/widget/image.h
@@ -2,6 +2,7 @@
 #include "ada.h"
 #include "mapgenerator.h"
 #include "styled_widget.h"
+#include "webgpu_helpers.h"
 
 using fetchImageCallback = std::function<void(void*, size_t)>;
