+++ packages/dear-imgui/cpp/include/widget/map_view.h
@@ -1,5 +1,6 @@
 #include "mapgenerator.h"
 #include "styled_widget.h"
+#include "webgpu_helpers.h"
 
 class MapView final : public StyledWidget {
 private:
