+++ packages/dear-imgui/cpp/src/main.cpp
@@ -7,6 +7,7 @@
 #include <emscripten/bind.h>
 #include <nlohmann/json.hpp>
 
+#include "color_helpers.h"
 #include "glwasm.cpp"
 #include "reactimgui.h"
